use radial chart component
