+++ components/radial-chart/types.ts
@@ -1,3 +1,4 @@
 export interface RadialChartProps {
   highlightedCategory: string;
+  categoriesPercentages: unknown;
 }
